ORION_github/desktop/requirements.txt
# Desktop (EXE) extra requirements
# pip install -r requirements.txt --user
pywebview>=4.4.1
pyinstaller>=6.0.0

# Shared requirements (mismos que web)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
torch>=2.6.0
transformers>=4.57.0
accelerate>=1.12.0
safetensors>=0.7.0
huggingface-hub>=0.36.0
requests>=2.32.0
httpx>=0.28.0
sounddevice>=0.5.0
soundfile>=0.13.0
numpy>=2.0.0
realtimestt>=0.3.0
PyAudio>=0.2.14
python-kasa>=0.10.0
duckduckgo-search>=8.0.0
psutil>=7.0.0
python-dotenv>=1.0.1
